[#36] hotfix: 회원가입 유효성검사 픽스 & 미니 프로필 로그아웃 상태추가 & apiSERVER 도메인변경
