+++ electron_frontend/src/components/login/LoginForm.tsx
@@ -12,8 +12,9 @@ const onSubmit = async (
   setErrorMessage: React.Dispatch<React.SetStateAction<string>>
 ) => {
   try {
-    const response = await loginRequest(data.username, data.password);
-    console.log(response);
+    await loginRequest(data.username, data.password).then((res) => {
+      console.log(res.headers);
+    });
     setErrorMessage("");
   } catch (error) {
     message.error("Login failed Try again later");
